Refactoring ErrorSummary component
