+++ src/registration-form/Error-Summary.js
@@ -1,7 +1,8 @@
+import React, { useRef, useEffect } from 'react';
 import styled from '@emotion/styled';
 
-export default styled.div`
-  max-width: 40em;
+const ErrorSummary = styled.div`
+  max-width: 41em;
   padding: 16px;
   margin-bottom: 20px;
   border: 4px solid #b00f1f;
@@ -30,3 +31,29 @@ export default styled.div`
     font-weight: 700;
   }
 `;
+
+export default ({ errors }) => {
+  const refContainer = useRef(null);
+
+  useEffect(() => {
+    // focus error summary element if inputted values are invalid
+    if (refContainer.current) refContainer.current.focus();
+  }, [errors])
+
+  return (
+    <>
+      {errors.count() > 0 &&
+        <ErrorSummary tabIndex="-1" ref={refContainer}>
+          <h2>問題があります</h2>
+          <ul>
+            {errors.all().map((error, i) => (
+              <li key={i}>
+                <a href={`#${error.key}`}>{error.message}</a>
+              </li>
+            ))}
+          </ul>
+        </ErrorSummary>
+      }
+    </>
+  )
+}
